requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
# Editable Git install with no remote (model-server==1.0)
-e /home/pag/Development/model_server
numpy==1.18.0
PyYAML==5.2
torch==1.3.1
Werkzeug==0.16.0
